backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
requests>=2.31.0
dnspython>=2.4.0
ipwhois>=1.2.0
python-whois>=0.8.0
cryptography>=41.0.0
